+++ CidadeInteligente.Application/Commands/UpdateUser/UpdateUserCommand.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace CidadeInteligente.Application.Commands.UpdateUser;
 
-public record UpdateUserCommand(long UserId, long CourseId, string Name, string Email, Role Role) : IRequest<Unit?> { }
+public record UpdateUserCommand(int UserId, int CourseId, string Name, string Email, Role Role) : IRequest<Unit?> { }
