+++ CidadeInteligente.Core/Entities/Area.cs
@@ -2,11 +2,11 @@
 
 public class Area
 {
-    public long AreaId { get; private set; }
+    public int AreaId { get; private set; }
     public string Description { get; private set; }
     public virtual ICollection<Project> Projects { get; private set; } = [];
 
-    public Area(long areaId, string description)
+    public Area(int areaId, string description)
     {
         AreaId = areaId;
         Description = description;
