+++ CidadeInteligente.Core/Repositories/ICourseRepository.cs
@@ -7,6 +7,6 @@ public interface ICourseRepository
     Task AddAsync(Course course);
     void Delete(Course course);
     Task<List<Course>> GetAllAsync();
-    Task<Course?> GetByIdAsync(long courseId, bool tracking = false);
-    Task<bool> HaveProjectsAsync(long courseId);
+    Task<Course?> GetByIdAsync(int courseId, bool tracking = false);
+    Task<bool> HaveProjectsAsync(int courseId);
 }
