+++ CidadeInteligente.Mvc/Requests/v1/CreateUserRequest.cs
@@ -2,4 +2,4 @@
 
 namespace CidadeInteligente.Mvc.Requests.v1;
 
-public record CreateUserRequest(long CourseId, string Name, string Email, string Password, Role Role);
+public record CreateUserRequest(int CourseId, string Name, string Email, string Password, Role Role);
