+++ CidadeInteligente.Mvc/Requests/v1/UpdateProjectRequest.cs
@@ -1,13 +1,13 @@
 ﻿namespace CidadeInteligente.Mvc.Requests.v1;
 
 public record UpdateProjectRequest(string Title,
-    long AreaId,
-    long CourseId,
+    int AreaId,
+    int CourseId,
     string? Description,
     DateOnly StartedAt,
     DateOnly? FinishedAt,
-    IEnumerable<long> InvolvedUsers,
+    IEnumerable<int> InvolvedUsers,
     IEnumerable<UpdateProjectRequest.UpdateMediaRequest> Medias)
 {
-    public record UpdateMediaRequest(long? MediaId, string Title, string? Description, IFormFile File);
+    public record UpdateMediaRequest(int? MediaId, string Title, string? Description, IFormFile File);
 }
